GToken minting draft
Minting of GTokens is based on the amount of the base pair equivalent to the
amount of LP tokens minted at that time.

I still have questions as if this is the right way to go about minting it.
If I go this way, then, how do users redeem GToken for any token in the
basket that they want? Will minting now suggest they are going long on
the base token? These and more questions are still on my mind.