GRatio per pair and GTokens at Initial Liquidity
Fees are envisaged to play a role in maitaining even spread of liquidity
across all pairs.

If the GToken supply for a pair increases, the fee ratio increases also
hence disincentivising users from further providing liquidity for that pair.
This could possibly create an arbitrage situation, hence the redistribution
of the base pair and g-pairs accross their pools.

The use of the initial liquidity tokens is not yet certain. I expect that
projects wanting to list with this SC will go through a governance vote
where they will provide the initial liquidity (maybe we use PD??), if they
get vetoed they loose a part of the initial liquidity, if it passes
the GTokens minted will be used for what? Maybe they will be added to the
icentive pool for those participating in governance by locking up their
GTokens.

The amount of LPTokens minted per pair is now being stored along side the
amount of GTokens minted for that pair. These two values are envisaged to
be used to calculate the amount of LPTokens a fraction of GToken can liquidate
at any time from the pool of this pair

Yet to implement should be a risk management means that prevents massive
liquidation of a pool per block. Maybe a max percentage per period (blocks,
timestamp range, or epochs??). This percentage could take the form of
```math
( 1 / n ) * 100
```
Where `n` is the number of pools actively supported by the protocol.

Another risk management means is one that should affect a defunct
pool, where an emergency halt will cause spread of loss to all GToken
holders by taken a perceive percent loss out of their GToken they submit
for redemptioni