Implemented Minting GToken with base pair
It feels like this implementation brings optimisation funds usage.
For e.g, the `GTokens` minted in the `mint` test increased and the base pair
dust on the contract reduced significantly.

I am still wondering if slippage should be handled by the contract, or user
defined. If we allow user defined, I think it will be fine since all the dust
generated from each mint are kept by the contract.

Also implementation of slippage on fee ratio might be a good way to go.
Since, as above, all dusts are kept by the contract, the user might be
disincentived to effect any malicious activity that this might cause.

Future implementation on the use of dust should include a silent minting of
GTokens with them. Subsequently, dust GTokens should be used to buy base
pair tokens and used to reward governance participants. Maybe the base
pair reserve used for this can be split into 360 parts, each added to the
governance reward pool each epoch.
The convertion and reward pool top up should be in one Tx since we will
want to avoid any confusion from dust base pair and that for governance
reward pool